moved TeamMeeting in meeting package
